fetching user relays and triggering sync for first user
