+++ server/utils/negentropy/checkIfUserExistsOnRelay.go
@@ -7,12 +7,14 @@ import (
 	"time"
 
 	"github.com/gorilla/websocket"
+
+	nostr "grain/server/types"
 )
 
 const WebSocketTimeout = 2 * time.Second // Timeout for WebSocket responses
 
 // CheckIfUserExistsOnRelay checks if a user exists on the relay by their pubkey.
-func CheckIfUserExistsOnRelay(pubKey string, relays []string) (bool, error) {
+func CheckIfUserExistsOnRelay(pubKey, eventID string, relays []string) (bool, error) {
 	for _, url := range relays {
 		// Connect to the relay's WebSocket
 		conn, _, err := websocket.DefaultDialer.Dial(url, nil)
@@ -27,12 +29,8 @@ func CheckIfUserExistsOnRelay(pubKey string, relays []string) (bool, error) {
 			Authors: []string{pubKey}, // Filter by the author (pubkey)
 		}
 
-		subID := "sub_check_user" // Unique subscription identifier
-		subRequest := []interface{}{
-			"REQ",
-			subID,
-			filter,
-		}
+		subID := "sub_check_user"
+		subRequest := []interface{}{"REQ", subID, filter}
 
 		requestJSON, err := json.Marshal(subRequest)
 		if err != nil {
@@ -40,22 +38,17 @@ func CheckIfUserExistsOnRelay(pubKey string, relays []string) (bool, error) {
 			return false, err
 		}
 
-		// Send the subscription request
 		if err := conn.WriteMessage(websocket.TextMessage, requestJSON); err != nil {
 			log.Printf("Failed to send subscription request: %v\n", err)
 			return false, err
 		}
 
-		// Channels for response handling
 		msgChan := make(chan []byte)
 		errChan := make(chan error)
-		done := make(chan struct{})
-		eventReceived := false // Track if any events are received
-		loggedOnce := false    // Avoid logging multiple times
+		eventCount := 0
+		isNewUser := true
 
-		// Goroutine to listen for WebSocket responses
 		go func() {
-			defer close(done) // Signal that the goroutine is finished
 			for {
 				_, message, err := conn.ReadMessage()
 				if err != nil {
@@ -66,83 +59,56 @@ func CheckIfUserExistsOnRelay(pubKey string, relays []string) (bool, error) {
 			}
 		}()
 
-		// Process messages until EOSE is received
 		for {
 			select {
 			case message := <-msgChan:
 				var response []interface{}
 				if err := json.Unmarshal(message, &response); err != nil {
 					log.Printf("Failed to unmarshal response: %v\n", err)
-					return false, err
+					continue
 				}
 
-				// Process response
-				if len(response) > 0 {
-					switch response[0] {
-					case "EVENT":
-						eventReceived = true
-						if !loggedOnce {
-							log.Printf("User exists: Found events for pubkey %s\n", pubKey)
-							loggedOnce = true
-						}
-
-					case "EOSE":
-						log.Printf("End of subscription signal received for pubkey %s\n", pubKey)
-
-						// Send CLOSE message
-						closeRequest := []interface{}{"CLOSE", subID}
-						closeJSON, err := json.Marshal(closeRequest)
-						if err != nil {
-							log.Printf("Failed to marshal CLOSE message: %v\n", err)
-							return false, err
-						}
-
-						if err := conn.WriteMessage(websocket.TextMessage, closeJSON); err != nil {
-							log.Printf("Failed to send CLOSE message: %v\n", err)
-							return false, err
-						}
-
-						// Wait for CLOSED response
-						select {
-						case closedMsg := <-msgChan:
-							var closedResponse []interface{}
-							if err := json.Unmarshal(closedMsg, &closedResponse); err != nil {
-								log.Printf("Failed to unmarshal CLOSED response: %v\n", err)
-								return false, err
-							}
-
-							if len(closedResponse) > 0 && closedResponse[0] == "CLOSED" {
-								log.Printf("Subscription closed successfully: %s\n", subID)
-								return !eventReceived, nil // Return true if no events were received (new user)
-							}
-
-						case err := <-errChan:
-							log.Printf("Error waiting for CLOSED response: %v\n", err)
-							return false, err
-
-						case <-time.After(WebSocketTimeout):
-							log.Printf("Timeout waiting for CLOSED response for subscription %s\n", subID)
-							return false, nil
-						}
-
-					default:
-						log.Printf("Unexpected response: %v\n", response)
+				if len(response) > 0 && response[0] == "EVENT" {
+					// Parse the event
+					eventData, ok := response[2].(map[string]interface{})
+					if !ok {
+						log.Printf("Unexpected event data type: %T\n", response[2])
+						continue
 					}
+
+					// Extract the event
+					eventBytes, err := json.Marshal(eventData)
+					if err != nil {
+						log.Printf("Failed to marshal event data: %v\n", err)
+						continue
+					}
+
+					var event nostr.Event
+					if err := json.Unmarshal(eventBytes, &event); err != nil {
+						log.Printf("Failed to unmarshal event: %v\n", err)
+						continue
+					}
+
+					// Skip the current event being processed
+					if event.ID == eventID {
+						continue
+					}
+
+					eventCount++
+					isNewUser = false
 				}
 
-			case err := <-errChan:
-				if err.Error() == "EOF" {
-					log.Printf("Connection closed cleanly by remote host.")
-					return !eventReceived, nil // Gracefully handle EOF
+				if len(response) > 0 && response[0] == "EOSE" {
+					return isNewUser, nil
 				}
-				log.Printf("Error reading WebSocket message: %v\n", err)
-				return false, err
-
 			case <-time.After(WebSocketTimeout):
-				log.Printf("WebSocket response timeout for pubkey %s\n", pubKey)
-				return false, nil
+				log.Printf("WebSocket timeout while checking user existence for pubkey: %s\n", pubKey)
+				return isNewUser, nil
+			case err := <-errChan:
+				log.Printf("Error reading from WebSocket: %v\n", err)
+				return false, err
 			}
 		}
 	}
-	return true, nil // Default to new user if no relays respond
+	return true, nil
 }
